# Conflicts:
#	src/main/java/com/microsoft/graph/models/generated/BaseActivityHistoryItem.java

+++ src/main/java/com/microsoft/graph/concurrency/ChunkedUploadProvider.java
@@ -82,7 +82,7 @@ public class ChunkedUploadProvider<UploadType> {
     /**
      * The stream size
      */
-    private final int streamSize;
+    private final long streamSize;
 
     /**
      * The upload response handler
@@ -92,7 +92,7 @@ public class ChunkedUploadProvider<UploadType> {
     /**
      * The counter for how many bytes have been read from input stream
      */
-    private int readSoFar;
+    private long readSoFar;
 
     /**
      * Creates the ChunkedUploadProvider
@@ -106,7 +106,7 @@ public class ChunkedUploadProvider<UploadType> {
     public ChunkedUploadProvider(final UploadSession uploadSession,
                                  final IGraphServiceClient client,
                                  final InputStream inputStream,
-                                 final int streamSize,
+                                 final long streamSize,
                                  final Class<UploadType> uploadTypeClass) {
         if (uploadSession == null) {
             throw new InvalidParameterException("Upload session is null.");
